% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simp_lin_cpp}
\alias{simp_lin_cpp}
\title{Efficiently fit a simple linear regression in C++}
\usage{
simp_lin_cpp(x, y)
}
\arguments{
\item{x}{Vector of covariate observations}

\item{y}{Vector of response observations}
}
\value{
An Rcpp List object containing model elements.
}
\description{
Efficiently fit a simple linear regression in C++
}
