% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{simp_lin_R}
\alias{simp_lin_R}
\title{Fit simple linear regression model}
\usage{
simp_lin_R(x, y)
}
\arguments{
\item{x}{Vector of covariate observations}

\item{y}{Vector of response observations}
}
\value{
List of model components
}
\description{
Fit simple linear regression model
}
