% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearmodel.R
\name{simulate}
\alias{simulate}
\title{Simulate a dataset and run simple linear regression}
\usage{
simulate(n, method = "lm", id = 0, outputprefix = "run_", writeout = TRUE)
}
\arguments{
\item{n}{"Number of observations to simulate"}

\item{method}{"Which regression function to use, either 'lm' or 'simplin'}

\item{id}{"run id number}

\item{outputprefix}{"prefix of output file name"}

\item{writeout}{"Whether or not to write a summary of the results to a csv file"}
}
\description{
Simulate a dataset and run simple linear regression
}
